+++ Conjugal/Affixing/Affixation.cs
@@ -1,5 +1,6 @@
 using System;
 using System.ComponentModel;
+using System.Text;
 
 using JetBrains.Annotations;
 
@@ -14,13 +15,13 @@ namespace FowlFever.Conjugal.Affixing {
     [PublicAPI]
     public readonly struct Affixation : IAffixed {
         /// <inheritdoc />
-        public string Stem { get; }
+        public string? Stem { get; init; }
         /// <inheritdoc />
-        public string BoundMorpheme { get; }
+        public string? BoundMorpheme { get; init; }
         /// <inheritdoc />
-        public string Joiner { get; }
+        public string? Joiner { get; init; }
         /// <inheritdoc />
-        public AffixFlavor AffixFlavor { get; }
+        public AffixFlavor AffixFlavor { get; init; }
         /// <summary>
         /// An extra, optional <see cref="int"/> used by some <see cref="AffixFlavor"/> types:
         /// <ul>
@@ -31,56 +32,19 @@ namespace FowlFever.Conjugal.Affixing {
         /// <br/>
         /// For other <see cref="AffixFlavor"/> types, <see cref="Index"/> <b>must not be set</b>.
         /// </summary>
-        internal int Index => AffixFlavor.GetIndex(_index);
-        private readonly int? _index;
-
-        /// <summary>
-        /// Constructs a new <see cref="Affixation"/> with the given parameters.
-        /// </summary>
-        /// <remarks>Please construct <see cref="Affixation"/> instances via the static factory methods: <see cref="Suffixation(string,string,string)"/>, <see cref="Ambifixation(string,string,string)"/>, etc.</remarks>
-        /// <exception cref="Index">If an <see cref="ArgumentException"/> is needed but missing <b>or</b> missing but needed.</exception>
-        internal Affixation(
-            AffixFlavor affixFlavor,
-            string?     stem,
-            string?     boundMorpheme,
-            int?        index,
-            string?     joiner = ""
-        ) {
-            if (AffixFlavorExtensions.RequiresIndex(affixFlavor) != index.HasValue) {
-                var problem = index.HasValue ? $"must not have an {nameof(index)} specified" : $"requires you to specify an {nameof(index)}";
-                throw new ArgumentException($"{nameof(global::FowlFever.Conjugal.Affixing.AffixFlavor)} {nameof(affixFlavor)} {affixFlavor} {problem} when constructing an {nameof(Affixation)}!");
-            }
-
-            Stem          = stem;
-            BoundMorpheme = boundMorpheme;
-            Joiner        = joiner;
-            AffixFlavor   = affixFlavor;
-            _index        = index;
+        internal Index Index {
+            get => AffixFlavor.ValidateIndex(_index);
+            init => _index = AffixFlavor.ValidateIndex(value);
         }
 
-        internal Affixation(
-            AffixFlavor affixFlavor,
-            string?     stem,
-            string?     boundMorpheme,
-            string?     joiner = ""
-        ) : this(
-            affixFlavor,
-            stem,
-            boundMorpheme,
-            default,
-            joiner
-        ) { }
+        /// <inheritdoc />
+        public Affixation(AffixFlavor flavor) : this() {
+            AffixFlavor = flavor;
+        }
 
-        internal Affixation(
-            IAffix  affix,
-            string? stem
-        ) : this(
-            affix.AffixFlavor,
-            stem,
-            affix.BoundMorpheme,
-            default,
-            affix.Joiner
-        ) { }
+        private readonly Index _index;
+
+        internal const string DefaultJoiner = "";
 
         /// <summary>
         /// Constructs a new <see cref="Affixation"/> instance with a <see cref="Affixing.AffixFlavor.Suffix"/>.
@@ -89,8 +53,13 @@ namespace FowlFever.Conjugal.Affixing {
         /// <param name="suffix">the <see cref="BoundMorpheme"/> appended to the <see cref="Stem"/></param>
         /// <param name="joiner">the <see cref="Joiner"/> interposed betwixt the <see cref="Stem"/> and <see cref="BoundMorpheme"/>. Defaults to <see cref="string.Empty">""</see>.</param>
         /// <returns>a new <see cref="Affixation"/></returns>
-        public static Affixation Suffixation(string? stem, string? suffix, string? joiner = "") {
-            return new Affixation(AffixFlavor.Suffix, stem, suffix, joiner);
+        public static Affixation Suffixation(string? stem, string? suffix, string? joiner = DefaultJoiner) {
+            return new Affixation {
+                AffixFlavor   = AffixFlavor.Suffix,
+                Stem          = stem,
+                BoundMorpheme = suffix,
+                Joiner        = joiner,
+            };
         }
 
         /// <summary>
@@ -100,8 +69,13 @@ namespace FowlFever.Conjugal.Affixing {
         /// <param name="prefix">the <see cref="BoundMorpheme"/> appended to the <see cref="Stem"/></param>
         /// <param name="joiner">the <see cref="Joiner"/> interposed betwixt the <see cref="Stem"/> and <see cref="BoundMorpheme"/></param>
         /// <returns>a new <see cref="Affixation"/></returns>
-        public static Affixation Prefixation(string? stem, string? prefix, string? joiner = "") {
-            return new Affixation(AffixFlavor.Prefix, stem, prefix, joiner);
+        public static Affixation Prefixation(string? stem, string? prefix, string? joiner = DefaultJoiner) {
+            return new Affixation {
+                AffixFlavor   = AffixFlavor.Prefix,
+                Stem          = stem,
+                BoundMorpheme = prefix,
+                Joiner        = joiner,
+            };
         }
 
         /// <summary>
@@ -112,8 +86,14 @@ namespace FowlFever.Conjugal.Affixing {
         /// <param name="index">the <see cref="string"/> index where the <see cref="string.Insert"/>ion will take place</param>
         /// <param name="joiner">the <see cref="Joiner"/> interposed betwixt the <see cref="Stem"/> and <see cref="BoundMorpheme"/>. Defaults to <see cref="string.Empty">""</see>.</param>
         /// <returns>a new <see cref="Affixation"/></returns>
-        public static Affixation Infixation(string? stem, string? infix, int index, string? joiner) {
-            return new Affixation(AffixFlavor.Infix, stem, infix, index, joiner);
+        public static Affixation Infixation(string? stem, string? infix, int index, string? joiner = DefaultJoiner) {
+            return new Affixation {
+                AffixFlavor   = AffixFlavor.Infix,
+                Stem          = stem,
+                BoundMorpheme = infix,
+                Index         = index,
+                Joiner        = joiner,
+            };
         }
 
         /// <summary>
@@ -131,8 +111,14 @@ namespace FowlFever.Conjugal.Affixing {
         /// <param name="suffix">the part of the <see cref="BoundMorpheme"/> that will <b>succede</b> the <see cref="Stem"/>(i.e. "en" in "embooben")</param>
         /// <param name="joiner">the <see cref="Joiner"/> interposed betwixt the <see cref="BoundMorpheme"/> and the <see cref="Stem"/>. Defaults to <see cref="string.Empty">""</see></param>
         /// <returns>a new <see cref="Affixation"/></returns>
-        public static Affixation Circumfixation(string? stem, string? prefix, string? suffix, string? joiner) {
-            return new Affixation(AffixFlavor.Circumfix, stem, prefix + suffix, prefix?.Length, joiner);
+        public static Affixation Circumfixation(string? stem, string? prefix, string? suffix, string? joiner = DefaultJoiner) {
+            return new Affixation() {
+                AffixFlavor   = AffixFlavor.Circumfix,
+                Stem          = stem,
+                BoundMorpheme = prefix + suffix,
+                Index         = prefix?.Length ?? 0,
+                Joiner        = joiner,
+            };
         }
 
         /// <summary>
@@ -143,37 +129,57 @@ namespace FowlFever.Conjugal.Affixing {
         /// <param name="ambifix">the <see cref="BoundMorpheme"/> that will appear on <b>BOTH SIDES</b> of the <see cref="Stem"/></param>
         /// <param name="joiner">the <see cref="Joiner"/> interposed betwixt the <see cref="BoundMorpheme"/> and the <see cref="Stem"/>. Defaults to <see cref="string.Empty">""</see></param>
         /// <returns></returns>
-        public static Affixation Ambifixation(string stem, string ambifix, string joiner) {
-            return new Affixation(AffixFlavor.Ambifix, stem, ambifix, joiner);
+        public static Affixation Ambifixation(string stem, string ambifix, string? joiner = DefaultJoiner) {
+            return new Affixation {
+                AffixFlavor   = AffixFlavor.Ambifix,
+                Stem          = stem,
+                BoundMorpheme = ambifix,
+                Joiner        = joiner,
+            };
         }
-
-        /// <summary>
-        /// The prefixial <see cref="BoundMorpheme"/> for use with <see cref="Circumfixation(string,string,string,string)"/>.
-        /// </summary>
-        /// <remarks>Fun fact: "circum" is a <a href="https://en.wikipedia.org/wiki/Libfix">libfix</a>!</remarks>
-        private string CircumPrefix => BoundMorpheme?.Substring(0, Index);
-
-        /// <summary>
-        /// The suffixial <see cref="BoundMorpheme"/> for use with <see cref="Circumfixation(string,string,string,string)"/>.
-        /// </summary>
-        /// <remarks>Fun fact: "circum" is a <a href="https://en.wikipedia.org/wiki/Libfix">libfix</a>!</remarks>
-        private string CircumSuffix => BoundMorpheme?.Substring(Index, BoundMorpheme.Length);
 
         /// <returns>the final result of this <see cref="Affixation"/></returns>
         /// <exception cref="NotImplementedException">if this <see cref="AffixFlavor"/> hasn't been implemented</exception>
         /// <exception cref="InvalidEnumArgumentException">if this <see cref="AffixFlavor"/> is unknown</exception>
         public string Render() {
+            return this switch {
+                { Stem.Length: > 0, BoundMorpheme.Length: > 0 } => _build(),
+                { Stem: not null }                              => Stem,
+                _                                               => "",
+            };
+        }
+
+        private string _build() {
             return AffixFlavor switch {
                 AffixFlavor.Prefix    => $"{BoundMorpheme}{Joiner}{Stem}",
                 AffixFlavor.Suffix    => $"{Stem}{Joiner}{BoundMorpheme}",
-                AffixFlavor.Infix     => Stem?.Insert(Index, BoundMorpheme.Ambifix(Joiner)) ?? "",
-                AffixFlavor.Circumfix => $"{CircumPrefix}{Joiner}{Stem}{Joiner}{CircumSuffix}",
-                AffixFlavor.Ambifix   => Stem.Circumfix(BoundMorpheme, BoundMorpheme, default),
+                AffixFlavor.Infix     => Stem?.Insert(Index.GetOffset(Stem.Length), $"{Joiner}{BoundMorpheme}{Joiner}") ?? "",
+                AffixFlavor.Circumfix => _BuildCircumfix(),
+                AffixFlavor.Ambifix   => $"{BoundMorpheme}{Joiner}{Stem}{Joiner}{BoundMorpheme}",
                 AffixFlavor.Duplifix  => throw new NotImplementedException($"{AffixFlavor} is not supported!"),
                 AffixFlavor.Transfix  => throw new NotImplementedException($"{AffixFlavor} is not supported!"),
                 AffixFlavor.Disfix    => throw new NotImplementedException($"{AffixFlavor} is not supported!"),
                 _                     => throw new InvalidEnumArgumentException(nameof(AffixFlavor), (int)AffixFlavor, typeof(AffixFlavor))
-            };
+            } ?? "";
+        }
+
+        private string _BuildCircumfix() {
+            var sb = new StringBuilder();
+
+            if (BoundMorpheme?[..Index] is { Length: > 0 } circumPrefix) {
+                sb.Append(circumPrefix);
+                sb.Append(Joiner);
+            }
+
+            sb.Append(Stem);
+
+            // ReSharper disable once InvertIf
+            if (BoundMorpheme?[Index..] is { Length: > 0 } circumSuffix) {
+                sb.Append(Joiner);
+                sb.Append(circumSuffix);
+            }
+
+            return sb.ToString();
         }
 
         /// <returns>individual properties of this <see cref="Affixation"/> for debugging</returns>