+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- `Affixation.Index` from `int` -> `System.Index`
+
+### Fixed
+- `Joiner`s should never be left dangling
+
 ## [1.0.0] - 2021-12-08
 
 ### Added